+++ packages/component/src/Styles/defaultStyleOptions.js
@@ -99,7 +99,7 @@ const DEFAULT_OPTIONS = {
 
   // Timestamp
   timestampColor: DEFAULT_SUBTLE,
-  timestampConfig: 'relative', // 'absolute'
+  timestampFormat: 'relative', // 'absolute'
 
   // Transcript overlay buttons (e.g. carousel and suggested action flippers, scroll to bottom, etc.)
   transcriptOverlayButtonBackground: 'rgba(0, 0, 0, .6)',